python_backend/requirements.txt
openai-whisper==20231117
flask==2.3.3
flask-cors==4.0.0
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3
scipy==1.11.3
librosa==0.10.1
pydub==0.25.1
python-multipart==0.0.6
uvicorn==0.23.2
fastapi==0.104.1